requirements.txt
numpy>=1.24.3
pandas>=2.0.2
scikit-learn>=1.2.2
python-dotenv>=1.0.0
nltk>=3.8.1
requests>=2.31.0
